# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Meu Auto API
  version: "1.0.0"
  description: |
    API do Meu Auto — acompanhamento da vida útil de um veículo para o dono pessoa física
    no Brasil.

    **Convenções**

    - Datas civis (dia de uma manutenção, vencimento de IPVA) são `YYYY-MM-DD`, sem hora e
      sem fuso. Timestamps de auditoria (`created_at`, `updated_at`) são RFC 3339.
    - Dinheiro é sempre **inteiro em centavos** (`*_cents`). Nunca float, nunca string.
    - Distância é sempre inteiro em quilômetros.
    - Mensagens de erro são pt-BR e podem ser exibidas ao usuário. O `code` é estável e em
      inglês — é nele que o cliente decide comportamento, nunca na mensagem.
    - `POST` de recurso aceita `id` gerado pelo cliente (UUIDv7). Reenviar a mesma
      requisição é idempotente: retorna `200` com o recurso já criado em vez de `201`.
    - Listas paginadas usam cursor opaco (`next_cursor`), nunca offset. `next_cursor: null`
      é a última página. O cursor é emitido pelo servidor — não construa um.
  contact:
    name: Meu Auto
  license:
    name: Proprietary
    url: https://meuauto.com.br
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://{host}
    description: Ambiente hospedado
    variables:
      host:
        default: api.meuauto.com.br
...
security:
  - bearerAuth: []
paths:
...
  /v1/vehicles/{vehicleId}/obligations:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      operationId: listObligations
      tags: [prazos]
      summary: IPVA e licenciamento
      parameters:
        - name: kind
          in: query
          schema: { $ref: "#/components/schemas/ObligationKind" }
      responses:
        "200":
          description: Os débitos do veículo, vencimento mais recente primeiro
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Obligation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    post:
      operationId: createObligation
      tags: [prazos]
      summary: Registrar IPVA ou licenciamento
      description: Um por tipo por ano de referência — um segundo é duplicata, não um segundo tributo.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateObligationRequest" }
      responses:
        "201":
          description: Débito registrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Obligation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/obligations/{obligationId}:
    parameters:
      - name: obligationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateObligation
      tags: [prazos]
      summary: Atualizar débito
      description: |
        Para desfazer um pagamento lançado por engano, envie `clear_payment: true` — sem
        dados de pagamento junto.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateObligationRequest" }
      responses:
        "200":
          description: Débito atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Obligation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteObligation
      tags: [prazos]
      summary: Apagar débito
      responses:
        "204": { description: Débito apagado }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/vehicles/{vehicleId}/seguros:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      operationId: listSeguros
      tags: [prazos]
      summary: Apólices de seguro
      responses:
        "200":
          description: As apólices, vigência mais recente primeiro
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Seguro" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createSeguro
      tags: [prazos]
      summary: Registrar apólice
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSeguroRequest" }
      responses:
        "201":
          description: Apólice registrada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Seguro" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/seguros/{seguroId}:
    parameters:
      - name: seguroId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateSeguro
      tags: [prazos]
      summary: Atualizar apólice
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSeguroRequest" }
      responses:
        "200":
          description: Apólice atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Seguro" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteSeguro
      tags: [prazos]
      summary: Apagar apólice
      responses:
        "204": { description: Apólice apagada }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- telas
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de acesso obtido em `/v1/auth/login`, `/register` ou `/refresh`. Vale 15
        minutos; renove com o refresh token antes de expirar.
  parameters:
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Autenticação necessária, ou sessão inválida/expirada
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        Recurso não encontrado.

        Também é a resposta para um recurso que existe mas não pertence ao usuário: dizer
        "você não pode ver isto" confirmaria que ele existe.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflito com um recurso existente
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: |
        Dados inválidos. `details.fields` traz **todos** os campos rejeitados de uma vez,
        com a chave igual à do JSON enviado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: validation_failed
              message: Não foi possível criar a conta.
              details:
                fields:
                  email: Informe um e-mail válido.
                  password: A senha deve ter pelo menos 8 caracteres.
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Mensagem em pt-BR, exibível ao usuário.
            details:
              type: object
              additionalProperties: true
              description: |
                Dados estruturados específicos do erro. Em `validation_failed` traz
                `fields`; em `odometer_rollback`, o registro vizinho conflitante; em
                `internal`, o `request_id` para suporte.
    ErrorCode:
      type: string
      description: |
        Identificador estável do erro. **É nele que o cliente decide comportamento**, nunca
        na mensagem.

        Valores podem ser acrescentados no futuro — trate um código desconhecido como falha
        genérica, nunca como erro de parsing.
      enum:
        - validation_failed
        - unauthorized
        - forbidden
        - not_found
        - method_not_allowed
        - conflict
        - odometer_rollback
        - rate_limited
        - internal

    # ------------------------------------------------------------ enums de domínio
...
    ObligationKind:
      type: string
      enum: [ipva, licenciamento]
    ObligationStatus:
      type: string
      description: |
        `pago` prevalece sobre a data — IPVA pago em atraso está pago, não vencido. Vencendo
        **hoje** é `vence_em_breve`, não `vencido`: ainda há horas para pagar.
      enum: [pago, vencido, vence_em_breve, pendente]
    SeguroStatus:
      type: string
      description: |
        `futuro` é uma renovação já contratada que ainda não entrou em vigor. `vencido`
        significa que o carro está **sem cobertura**.
      enum: [futuro, vigente, vence_em_breve, vencido]
...
    CreateObligationRequest:
      type: object
      required: [kind, reference_year, due_on]
      properties:
        id: { type: string, format: uuid, nullable: true }
        kind: { $ref: "#/components/schemas/ObligationKind" }
        reference_year:
          type: integer
          format: int32
          description: Ano de referência do tributo. Teto é o ano atual + 1.
        due_on:
          type: string
          format: date
          description: Vencimento. **Pode** estar no futuro — normalmente está.
        amount_cents: { type: integer, format: int64, nullable: true, minimum: 0 }
        paid_on:
          type: string
          format: date
          nullable: true
          description: Data do pagamento. Não pode estar no futuro.
        paid_amount_cents: { type: integer, format: int64, nullable: true, minimum: 0 }
        notes: { type: string, nullable: true, maxLength: 500 }
    UpdateObligationRequest:
      type: object
      properties:
        due_on: { type: string, format: date, nullable: true }
        amount_cents: { type: integer, format: int64, nullable: true, minimum: 0 }
        paid_on: { type: string, format: date, nullable: true }
        paid_amount_cents: { type: integer, format: int64, nullable: true, minimum: 0 }
        notes: { type: string, nullable: true }
        clear_payment:
          type: boolean
          description: |
            Marca o débito como não pago, desfazendo um lançamento feito por engano. Precisa
            de flag própria porque `null` já significa "não altere". Não pode vir junto de
            dados de pagamento.
    Obligation:
      type: object
      required:
        [id, vehicle_id, kind, reference_year, due_on, status, remaining_days, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        vehicle_id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/ObligationKind" }
        reference_year: { type: integer, format: int32 }
        due_on: { type: string, format: date }
        amount_cents: { type: integer, format: int64, nullable: true }
        paid_on: { type: string, format: date, nullable: true }
        paid_amount_cents: { type: integer, format: int64, nullable: true }
        notes: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ObligationStatus" }
        remaining_days:
          type: integer
          description: |
            Dias até o vencimento; negativo quando já passou. Devolvido mesmo quando pago,
            para a tela poder dizer "pago com 3 dias de atraso".
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateSeguroRequest:
      type: object
      required: [insurer_name, starts_on, ends_on]
      properties:
        id: { type: string, format: uuid, nullable: true }
        insurer_name: { type: string, maxLength: 120 }
        policy_number: { type: string, nullable: true, maxLength: 120 }
        starts_on: { type: string, format: date }
        ends_on:
          type: string
          format: date
          description: Não pode ser anterior a `starts_on`.
        premium_cents: { type: integer, format: int64, nullable: true, minimum: 0 }
        emergency_phone:
          type: string
          nullable: true
          maxLength: 32
          description: |
            O 0800 da seguradora. Só o tamanho é validado — formato de telefone brasileiro
            varia demais, e rejeitar um número que o dono consegue discar seria pior que
            guardar um estranho.
        broker_name: { type: string, nullable: true, maxLength: 120 }
        broker_phone: { type: string, nullable: true, maxLength: 32 }
        notes: { type: string, nullable: true, maxLength: 500 }
    UpdateSeguroRequest:
      type: object
      properties:
        insurer_name: { type: string, nullable: true, maxLength: 120 }
        policy_number: { type: string, nullable: true }
        starts_on: { type: string, format: date, nullable: true }
        ends_on: { type: string, format: date, nullable: true }
        premium_cents: { type: integer, format: int64, nullable: true, minimum: 0 }
        emergency_phone: { type: string, nullable: true }
        broker_name: { type: string, nullable: true }
        broker_phone: { type: string, nullable: true }
        notes: { type: string, nullable: true }
    Seguro:
      type: object
      required:
        [id, vehicle_id, insurer_name, starts_on, ends_on, status, remaining_days, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        vehicle_id: { type: string, format: uuid }
        insurer_name: { type: string }
        policy_number: { type: string, nullable: true }
        starts_on: { type: string, format: date }
        ends_on: { type: string, format: date }
        premium_cents: { type: integer, format: int64, nullable: true }
        emergency_phone: { type: string, nullable: true }
        broker_name: { type: string, nullable: true }
        broker_phone: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/SeguroStatus" }
        remaining_days:
          type: integer
          description: |
            Dias até o fim da vigência; negativo quando já acabou. Para uma apólice `futuro`,
            conta até o **início**.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ------------------------------------------------------------ telas
...