# internal/obligation/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"time"

	"github.com/google/uuid"
)

type MaintenanceItem struct {
	ID                    uuid.UUID
	Slug                  string
	Name                  string
	Kind                  string
	VehicleType           string
	OwnerUserID           *uuid.UUID
	DefaultIntervalKm     *int32
	DefaultIntervalMonths *int32
	DefaultIntervalDays   *int32
	SuggestByDefault      bool
	IsActive              bool
	CreatedAt             time.Time
}

type MaintenancePlan struct {
	ID                uuid.UUID
	VehicleID         uuid.UUID
	MaintenanceItemID uuid.UUID
	IntervalKm        *int32
	IntervalMonths    *int32
	IntervalDays      *int32
	AlertKm           int32
	AlertDays         int32
	Origin            string
	IsActive          bool
	CreatedAt         time.Time
	UpdatedAt         time.Time
}

type MaintenanceRecord struct {
	ID               uuid.UUID
	VehicleID        uuid.UUID
	OccurredOn       time.Time
	MileageKm        int32
	Kind             string
	WorkshopName     *string
	TotalCostCents   int64
	Notes            *string
	RecordedByUserID *uuid.UUID
	CreatedAt        time.Time
	UpdatedAt        time.Time
	DeletedAt        *time.Time
}

type MaintenanceRecordItem struct {
	ID                  uuid.UUID
	MaintenanceRecordID uuid.UUID
	MaintenanceItemID   uuid.UUID
	Description         *string
	PartBrand           *string
	CostCents           *int64
	WarrantyMonths      *int32
	WarrantyKm          *int32
	CreatedAt           time.Time
}

type OdometerReading struct {
	ID                  uuid.UUID
	VehicleID           uuid.UUID
	MileageKm           int32
	OccurredOn          time.Time
	Source              string
	RecordedByUserID    *uuid.UUID
	Notes               *string
	CreatedAt           time.Time
	SourceMaintenanceID *uuid.UUID
}

type PasswordResetToken struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	TokenHash []byte
	ExpiresAt time.Time
	UsedAt    *time.Time
	CreatedAt time.Time
}

type RefreshToken struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	TokenHash  []byte
	ExpiresAt  time.Time
	RevokedAt  *time.Time
	ReplacedBy *uuid.UUID
	UserAgent  *string
	CreatedAt  time.Time
}

type Seguro struct {
	ID               uuid.UUID
	VehicleID        uuid.UUID
	InsurerName      string
	PolicyNumber     *string
	StartsOn         time.Time
	EndsOn           time.Time
	PremiumCents     *int64
	EmergencyPhone   *string
	BrokerName       *string
	BrokerPhone      *string
	Notes            *string
	RecordedByUserID *uuid.UUID
	CreatedAt        time.Time
	UpdatedAt        time.Time
}

type User struct {
	ID           uuid.UUID
	Email        string
	PasswordHash string
	Name         string
	CreatedAt    time.Time
	UpdatedAt    time.Time
}

type Vehicle struct {
	VehicleType      string
	ID               uuid.UUID
	Brand            string
	Model            string
	Version          *string
	ManufactureYear  *int32
	ModelYear        *int32
	Plate            *string
	Renavam          *string
	Chassis          *string
	FuelType         *string
	Color            *string
	Nickname         *string
	FipeCode         *string
	CurrentMileageKm int32
	CurrentMileageAt *time.Time
	CreatedAt        time.Time
	UpdatedAt        time.Time
	DeletedAt        *time.Time
}

type VehicleObligation struct {
	ID               uuid.UUID
	VehicleID        uuid.UUID
	Kind             string
	ReferenceYear    int32
	DueOn            time.Time
	AmountCents      *int64
	PaidOn           *time.Time
	PaidAmountCents  *int64
	Notes            *string
	RecordedByUserID *uuid.UUID
	CreatedAt        time.Time
	UpdatedAt        time.Time
}

type VehicleOwnership struct {
	ID        uuid.UUID
	VehicleID uuid.UUID
	UserID    uuid.UUID
	Role      string
	StartedOn time.Time
	EndedOn   *time.Time
	CreatedAt time.Time
}